gym require optimizations/fixes
