+++ gym/lib/gym/commands_generator.rb
@@ -1,12 +1,15 @@
-require "commander"
-require "fastlane_core"
+require 'commander'
+
+require 'fastlane_core/configuration/configuration'
+require_relative 'module'
+require_relative 'manager'
+require_relative 'options'
 
 HighLine.track_eof = false
 
 module Gym
   class CommandsGenerator
     include Commander::Methods
-    UI = FastlaneCore::UI
 
     def self.start
       new.run
