+++ gym/lib/gym/module.rb
@@ -1,0 +1,36 @@
+require 'fastlane_core/helper'
+require 'fastlane/boolean'
+require_relative 'detect_values'
+
+module Gym
+  class << self
+    attr_accessor :config
+
+    attr_accessor :project
+
+    attr_accessor :cache
+
+    def config=(value)
+      @config = value
+      DetectValues.set_additional_default_values
+      @cache = {}
+    end
+
+    def gymfile_name
+      "Gymfile"
+    end
+
+    def init_libs
+      # Import all the fixes
+      require 'gym/xcodebuild_fixes/generic_archive_fix'
+    end
+  end
+
+  Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
+  UI = FastlaneCore::UI
+  Boolean = Fastlane::Boolean
+  ROOT = Pathname.new(File.expand_path('../..', __FILE__))
+  DESCRIPTION = "Building your iOS apps has never been easier"
+
+  Gym.init_libs
+end
